requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.196
python-dotenv==1.0.0
Pillow==10.0.1